requirements.txt
pyautogui
keyboard
mss
Pillow
